Move demo apps to separate integration-test directory
