+++ src/integration-test/java/com/scylladb/alternator/demo/Demo4.java
@@ -1,4 +1,4 @@
-package com.scylladb.alternator.test;
+package com.scylladb.alternator.demo;
 
 import com.scylladb.alternator.AlternatorDynamoDbClient;
 import java.net.URI;
